Add torsion routines
