+++ src/Structure/SugarBuilder.cpp
@@ -2,8 +2,12 @@
 #include "../ArgList.h"
 #include "../CpptrajFile.h"
 #include "../CpptrajStdio.h"
+#include "../TorsionRoutines.h"
 
 using namespace Cpptraj::Structure;
+
+/** CONSTRUCTOR */
+SugarBuilder::SugarBuilder(int debugIn) : debug_(debugIn) {}
 
 /** Load reduced internal PDB to Glycam map. */
 void SugarBuilder::SetGlycamPdbResMap() {
@@ -403,3 +407,246 @@ const
   }
   return 0;
 }
+// -----------------------------------------------------------------------------
+// Torsion routines
+/// \return Position of given tgt atom in the array (if it is in the given array)
+static inline int AtomIdxInArray(std::vector<int> const& ChainAtoms, int tgt)
+{
+  for (std::vector<int>::const_iterator it = ChainAtoms.begin(); it != ChainAtoms.end(); ++it)
+    if (*it == tgt)
+      return (int)(it - ChainAtoms.begin());
+  return -1;
+}
+
+/** Determine torsion around the anomeric carbon. */
+int SugarBuilder::CalcAnomericTorsion(double& torsion,
+                                             int anomeric_atom, int ring_oxygen_atom,
+                                             int rnum,
+                                             Iarray const& RingAtoms,
+                                             Topology const& topIn, Frame const& frameIn)
+const
+{
+  if (debug_ > 0) {
+    mprintf("\t  Anomeric carbon             : %s\n", topIn.ResNameNumAtomNameNum(anomeric_atom).c_str());
+    mprintf("\t  Ring oxygen atom            : %s\n", topIn.ResNameNumAtomNameNum(ring_oxygen_atom).c_str());
+  }
+  int anomeric_atom_X = -1;
+  int anomeric_atom_C = -1;
+  // By definition the anomeric atom should be the first ring atom TODO catch size==1?
+  anomeric_atom_C = RingAtoms[1];
+  if (anomeric_atom_C == -1) {
+    mprinterr("Error: Next ring atom after anomeric C could not be identified.\n");
+    return 1;
+  }
+  if (debug_ > 0)
+    mprintf("\t  Anomeric C ring substituent : %s\n",
+            topIn.ResNameNumAtomNameNum(anomeric_atom_C).c_str());
+  // Get the substituent of the anomeric C (e.g. C1) that is a non-ring atom, non hydrogen 
+  for ( Atom::bond_iterator bat = topIn[anomeric_atom].bondbegin();
+                            bat != topIn[anomeric_atom].bondend();
+                          ++bat )
+  {
+    if ( *bat != ring_oxygen_atom &&
+         topIn[*bat].Element() != Atom::HYDROGEN &&
+         !AtomIsInArray(RingAtoms, *bat) )
+    {
+      if (anomeric_atom_X != -1) {
+        // If there are two non-ring, non-hydrogen substituents, prioritize
+        // the one that is part of this residue.
+        bool bat_in_res = (topIn[*bat].ResNum() == rnum);
+        bool X_in_res   = (topIn[anomeric_atom_X].ResNum() == rnum);
+        if ( (bat_in_res && X_in_res) || (!bat_in_res && !X_in_res) ) {
+          // Both in or both out of residue. Choose based on atomic number.
+          if (topIn[*bat].AtomicNumber() == topIn[anomeric_atom_X].AtomicNumber()) {
+            mprinterr("Error: Two potential substituents for anomeric carbon: %s and %s\n",
+                      topIn.ResNameNumAtomNameNum(*bat).c_str(),
+                      topIn.ResNameNumAtomNameNum(anomeric_atom_X).c_str());
+            return 1;
+          } else if (topIn[*bat].AtomicNumber() > topIn[anomeric_atom_X].AtomicNumber()) {
+            anomeric_atom_X = *bat;
+          }
+        } else if (bat_in_res) {
+          anomeric_atom_X = *bat;
+        }
+      } else
+        anomeric_atom_X = *bat;
+    }
+  }
+
+  if (anomeric_atom_X == -1) {
+    // If the Cx (C1 substituent, usually a different residue) index is
+    // not found this usually means missing inter-residue bond.
+    // Alternatively, this could be an isolated sugar missing an -OH
+    // group, so make this non-fatal.
+    mprintf("Warning: Anomeric C non-ring substituent could not be identified.\n");
+    //        "Warning: This can happen if the sugar is bonded to something that\n"
+    //        "Warning:  is missing, e.g. a -OH group. In that case the coordinates\n"
+    //        "Warning   for the missing atoms may need to be generated.\n");
+    return -1;
+  }
+  if (debug_ > 0)
+    mprintf("\t  Anomeric X substituent      : %s\n",
+            topIn.ResNameNumAtomNameNum(anomeric_atom_X).c_str());
+
+  torsion = Torsion( frameIn.XYZ(ring_oxygen_atom), frameIn.XYZ(anomeric_atom),
+                     frameIn.XYZ(anomeric_atom_C), frameIn.XYZ(anomeric_atom_X) );
+  if (debug_ > 0)
+    mprintf("DEBUG: Anomeric torsion %s-%s-%s-%s= %f\n",
+            *(topIn[ring_oxygen_atom].Name()),
+            *(topIn[anomeric_atom].Name()),
+            *(topIn[anomeric_atom_C].Name()),
+            *(topIn[anomeric_atom_X].Name()),
+            torsion * Constants::RADDEG);
+  return 0;
+}
+
+/** Determine torsion around anomeric reference carbon. */
+int SugarBuilder::CalcAnomericRefTorsion(double& torsion,
+                                                int ano_ref_atom, int ring_oxygen_atom,
+                                                int ring_end_atom, Iarray const& RingAtoms,
+                                                Topology const& topIn, Frame const& frameIn)
+const
+{
+  if (debug_ > 0)
+    mprintf("\t  Anomeric ref carbon                   : %s\n",
+            topIn.ResNameNumAtomNameNum(ano_ref_atom).c_str());
+  //      ano_ref_atom_Y
+  //           |
+  //      ano_ref_atom
+  //       |        |
+  // ano_ref_atom_0 ano_ref_atom_1
+  int ano_ref_atom_Y = -1;
+  int ano_ref_atom_0 = -1;
+  int ano_ref_atom_1 = -1;
+  // This will be the index of the anomeric atom in the RingAtoms array
+  int ar_index = -1;
+  // Find ring atom that precedes the anomeric reference atom TODO catch size==1?
+  for (unsigned int idx = 1; idx != RingAtoms.size(); idx++) {
+    if (RingAtoms[idx] == ano_ref_atom) {
+        ar_index = (int)idx;
+        ano_ref_atom_0 = RingAtoms[idx-1];
+        break;
+    }
+  }
+  if (ano_ref_atom_0 == -1) {
+    mprinterr("Error: Anomeric reference ring C previous ring atom could not be identified.\n");
+    return 1;
+  }
+  if (debug_ > 0)
+    mprintf("\t  Anomeric reference previous ring atom : %s\n",
+            topIn.ResNameNumAtomNameNum(ano_ref_atom_0).c_str());
+  // If the anomeric reference atom is the ring end atom then ano_ref_atom_1
+  // is the ring oxygen.
+  if (ano_ref_atom == ring_end_atom) {
+    ano_ref_atom_1 = ring_oxygen_atom;
+  } else {
+    // Anomeric reference atom is somewhere before the ring end atom.
+    ano_ref_atom_1 = RingAtoms[ar_index+1];
+  }
+  if (debug_ > 0)
+    mprintf("\t  Anomeric reference next atom          : %s\n",
+            topIn.ResNameNumAtomNameNum(ano_ref_atom_1).c_str());
+  // Get non-hydrogen substituent of anomeric ref (e.g. C5) that 
+  // is not part of the ring (e.g. C6).
+  for ( Atom::bond_iterator bat = topIn[ano_ref_atom].bondbegin();
+                            bat != topIn[ano_ref_atom].bondend();
+                          ++bat )
+  {
+    if ( *bat != ring_oxygen_atom &&
+         topIn[*bat].Element() != Atom::HYDROGEN &&
+         !AtomIsInArray(RingAtoms, *bat) )
+    {
+      if (ano_ref_atom_Y != -1) {
+        mprinterr("Error: Two potential non-ring substituents for anomeric ref: %s and %s\n",
+                  topIn.ResNameNumAtomNameNum(*bat).c_str(),
+                  topIn.ResNameNumAtomNameNum(ano_ref_atom_Y).c_str());
+        return 1;
+      }
+      ano_ref_atom_Y = *bat;
+    }
+  }
+  if (ano_ref_atom_Y == -1) {
+    mprinterr("Error: Anomeric reference Y substituent could not be identified.\n");
+    return 1;
+  }
+  if (debug_ > 0)
+    mprintf("\t  Anomeric reference substituent        : %s\n",
+            topIn.ResNameNumAtomNameNum(ano_ref_atom_Y).c_str());
+
+  torsion = Torsion( frameIn.XYZ(ano_ref_atom_0),   frameIn.XYZ(ano_ref_atom),
+                     frameIn.XYZ(ano_ref_atom_1),   frameIn.XYZ(ano_ref_atom_Y) );
+  if (debug_ > 0)
+    mprintf("DEBUG: Anomeric reference torsion %s-%s-%s-%s= %f\n",
+            *(topIn[ano_ref_atom_0].Name()),
+            *(topIn[ano_ref_atom].Name()),
+            *(topIn[ano_ref_atom_1].Name()),
+            *(topIn[ano_ref_atom_Y].Name()),
+            torsion * Constants::RADDEG);
+  return 0;
+}
+
+/** Determine torsion around the configurational carbon. 
+  * Calculate torsion around config. carbon C as:
+  *   C0-C-Z-C1
+  * where C0 is the carbon preceding C in the chain, C1 is the carbon
+  * after C in the chain, and Z is the non-hydrogen substituent of C
+  * with highest priority. Do it this way to be consistent with how
+  * CalcAnomericRefTorsion orders the atoms.
+  */
+int SugarBuilder::CalcConfigCarbonTorsion(double& torsion, int config_carbon,
+                                                 Iarray const& carbon_chain,
+                                                 Topology const& topIn,
+                                                 Frame const& frameIn)
+const
+{
+  int atom_c0 = -1;
+  int atom_c1 = -1;
+  int atom_z  = -1;
+  // Get c0 and c1
+  int c_idx = AtomIdxInArray(carbon_chain, config_carbon);
+  if (c_idx < 1) {
+    mprinterr("Error: Could not determine carbon before config. C '%s'\n",
+              topIn.ResNameNumAtomNameNum(config_carbon).c_str());
+    return 1;
+  }
+  atom_c0 = carbon_chain[c_idx-1];
+  if ((unsigned int)c_idx+1 >= carbon_chain.size()) {
+    mprinterr("Error: Could not determine carbon after config. C '%s'\n",
+              topIn.ResNameNumAtomNameNum(config_carbon).c_str());
+    return 1;
+  }
+  atom_c1 = carbon_chain[c_idx+1];
+
+  for (Atom::bond_iterator bat = topIn[config_carbon].bondbegin();
+                           bat != topIn[config_carbon].bondend(); ++bat)
+  {
+    if (topIn[*bat].Element() != Atom::HYDROGEN &&
+        !AtomIsInArray(carbon_chain, *bat))
+    {
+      if (atom_z == -1)
+        atom_z = *bat;
+      else if (topIn[*bat].AtomicNumber() > topIn[atom_z].AtomicNumber())
+        atom_z = *bat;
+    }
+  }
+  if (atom_z == -1) {
+    mprinterr("Error: Could not determine substituent for config. C '%s'\n",
+              topIn.ResNameNumAtomNameNum(config_carbon).c_str());
+    return 1;
+  }
+
+  torsion = Torsion( frameIn.XYZ(atom_c0),
+                     frameIn.XYZ(config_carbon),
+                     frameIn.XYZ(atom_z),
+                     frameIn.XYZ(atom_c1) );
+  if (debug_ > 0)
+    mprintf("DEBUG: Config. C torsion %s-%s-%s-%s= %f\n",
+            *(topIn[atom_c0].Name()),
+            *(topIn[config_carbon].Name()),
+            *(topIn[atom_z].Name()),
+            *(topIn[atom_c1].Name()),
+            torsion*Constants::RADDEG);
+  return 0;
+}
+
+
